test: split pure-logic request_headers splice tests from the handler test (#684)
The per-format splice helpers are pure logic — exercised via reflection with the
constructor skipped, they need no Router/JWT/DB. Extending AbstractHandlerTestCase
meant its setUp() skipped them whenever JWT_SECRET was unset, needlessly dropping
that coverage. Move the six splice tests into CalendarRequestHeadersSpliceTest
(PHPUnit\Framework\TestCase, no setup) and keep only the handle()-driven
end-to-end test in CalendarRequestHeadersCacheTest. No assertion changes.
